feat: reworked dataloader; now central class for all sub-dataloaders e.g. indicator_loader
